display -> print table
